Add files via upload
updated version of bond yield macro - now with Freq/Freq2/Redemption_Value and Convention as optional fields, (base of 2,2,100 and "30/360") and uses Crew Macro Ensure Fields to adjust for inclusion. Adjusts pricing to 100 based par inside of macro.


Checks for R Package jvrFinance's installation and installs if package is not present.